Read legacy reservation in 1.8 migrations.
