+++ src/main/scala/mesosphere/marathon/core/instance/Reservation.scala
@@ -159,45 +159,4 @@ object Reservation {
     Writes[Reservation.Id] { id => JsString(id.label) }
   )
   implicit val reservationFormat: OFormat[Reservation] = Json.format[Reservation]
-
-  /**
-    * Infer the reservation id for an instance.
-    *
-    * In older Marathon versions the reservation id was the run spec path and the uuid of the instance
-    * joined by a separator. Eg ephemeral tasks and persistent apps used it. This is expressed by
-    * [[mesosphere.marathon.core.instance.Reservation.LegacyId]].
-    *
-    * Later Marathon versions used the instance id, ie `<run spec path>.marathon-<uuid>`, as the
-    * reservation id. This is expressed by [[mesosphere.marathon.core.instance.Reservation.SimplifiedId]].
-    * Notice the extra "marathon-" in the id string.
-    *
-    * The reservation id was in all cases determined by the `appTask.taskId`. The app task is just the
-    * first task of the tasks map of an instance.
-    *
-    * This method is only used if the saved reservation does not include an id. This will be true for
-    * all instance from apps and pods started with Marathon 1.8.194-1590825ea and earlier. All apps
-    * and pods from later version will have a reservation id persisted.
-    *
-    * Future Marathon versions that only allow upgrades from Marathon 1.9 and later can drop the
-    * inference and should safely assume that all reservation have a persisted id.
-    *
-    * @param tasksMap All tasks of an instance.
-    * @param instanceId The id of the instance this reservation belongs to.
-    * @return The proper reservation id.
-    */
-  def inferReservationId(tasksMap: Map[Task.Id, Task], instanceId: Instance.Id): Reservation.Id = {
-    if (tasksMap.nonEmpty) {
-      val taskId = appTask(tasksMap).getOrElse(throw new IllegalStateException(s"No task in $instanceId")).taskId
-      taskId match {
-        case Task.LegacyId(runSpecId, separator: String, uuid) => LegacyId(runSpecId, separator, uuid)
-        case Task.LegacyResidentId(runSpecId, separator, uuid, _) => LegacyId(runSpecId, separator, uuid)
-        case Task.EphemeralTaskId(instanceId, _) => SimplifiedId(instanceId)
-        case Task.TaskIdWithIncarnation(instanceId, _, _) => SimplifiedId(instanceId)
-      }
-    } else {
-      SimplifiedId(instanceId)
-    }
-  }
-
-  def appTask(tasksMap: Map[Task.Id, Task]): Option[Task] = tasksMap.headOption.map(_._2)
 }
