+++ src/main/scala/mesosphere/marathon/storage/migration/Helper.scala
@@ -9,12 +9,17 @@ import akka.stream.Materializer
 import akka.stream.scaladsl.{Flow, Sink}
 import com.typesafe.scalalogging.StrictLogging
 import mesosphere.marathon.core.instance.Instance.Id
+import mesosphere.marathon.core.instance.Reservation.{LegacyId, SimplifiedId}
+import mesosphere.marathon.core.instance.{LocalVolumeId, Reservation}
 import mesosphere.marathon.core.storage.store.{IdResolver, PersistenceStore}
 import mesosphere.marathon.core.storage.store.impl.cache.{LazyCachingPersistenceStore, LazyVersionCachingPersistentStore, LoadTimeCachingPersistenceStore}
 import mesosphere.marathon.core.storage.store.impl.zk.{ZkId, ZkPersistenceStore, ZkSerialized}
+import mesosphere.marathon.core.task.Task
 import mesosphere.marathon.state.Instance
 import mesosphere.marathon.storage.repository.InstanceRepository
-import play.api.libs.json.{JsValue, Json}
+import play.api.libs.json._
+import play.api.libs.json.Reads._
+import play.api.libs.functional.syntax._
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -111,4 +116,63 @@ object InstanceMigration extends MaybeStore with StrictLogging {
       case ZkSerialized(byteString) =>
         Json.parse(byteString.utf8String)
     }
+
+  /**
+    * JSON reads format for reservations that have no id saved. See [[MigrationTo17]] for an example.
+    *
+    * @param tasksMap
+    * @param instanceId
+    * @return
+    */
+  def legacyReservationReads(tasksMap: Map[Task.Id, Task], instanceId: Id): Reads[Reservation] = {
+    (
+      (__ \ "volumeIds").read[Seq[LocalVolumeId]] ~
+      (__ \ "state").read[Reservation.State] ~
+      (__ \ "id").readNullable[Reservation.Id]
+    ) { (volumesIds, state, maybeId) =>
+        val reservationId = maybeId.getOrElse(inferReservationId(tasksMap, instanceId))
+        Reservation(volumesIds, state, reservationId)
+      }
+  }
+
+  /**
+    * Infer the reservation id for an instance.
+    *
+    * In older Marathon versions the reservation id was the run spec path and the uuid of the instance
+    * joined by a separator. Eg ephemeral tasks and persistent apps used it. This is expressed by
+    * [[mesosphere.marathon.core.instance.Reservation.LegacyId]].
+    *
+    * Later Marathon versions used the instance id, ie `<run spec path>.marathon-<uuid>`, as the
+    * reservation id. This is expressed by [[mesosphere.marathon.core.instance.Reservation.SimplifiedId]].
+    * Notice the extra "marathon-" in the id string.
+    *
+    * The reservation id was in all cases determined by the `appTask.taskId`. The app task is just the
+    * first task of the tasks map of an instance.
+    *
+    * This method is only used if the saved reservation does not include an id. This will be true for
+    * all instance from apps and pods started with Marathon 1.8.194-1590825ea and earlier. All apps
+    * and pods from later version will have a reservation id persisted.
+    *
+    * Future Marathon versions that only allow upgrades from Marathon 1.9 and later can drop the
+    * inference and should safely assume that all reservation have a persisted id.
+    *
+    * @param tasksMap All tasks of an instance.
+    * @param instanceId The id of the instance this reservation belongs to.
+    * @return The proper reservation id.
+    */
+  private[migration] def inferReservationId(tasksMap: Map[Task.Id, Task], instanceId: Id): Reservation.Id = {
+    if (tasksMap.nonEmpty) {
+      val taskId = appTask(tasksMap).getOrElse(throw new IllegalStateException(s"No task in $instanceId")).taskId
+      taskId match {
+        case Task.LegacyId(runSpecId, separator: String, uuid) => LegacyId(runSpecId, separator, uuid)
+        case Task.LegacyResidentId(runSpecId, separator, uuid, _) => LegacyId(runSpecId, separator, uuid)
+        case Task.EphemeralTaskId(instanceId, _) => SimplifiedId(instanceId)
+        case Task.TaskIdWithIncarnation(instanceId, _, _) => SimplifiedId(instanceId)
+      }
+    } else {
+      SimplifiedId(instanceId)
+    }
+  }
+
+  def appTask(tasksMap: Map[Task.Id, Task]): Option[Task] = tasksMap.headOption.map(_._2)
 }
