+++ src/main/scala/mesosphere/marathon/storage/migration/MigrationTo17.scala
@@ -89,17 +89,6 @@ object MigrationTo17 extends StrictLogging {
     }
   }
 
-  def legacyReservationReads(tasksMap: Map[Task.Id, Task], instanceId: Id): Reads[Reservation] = {
-    (
-      (__ \ "volumeIds").read[Seq[LocalVolumeId]] ~
-      (__ \ "state").read[Reservation.State] ~
-      (__ \ "id").readNullable[Reservation.Id]
-    ) { (volumesIds, state, maybeId) =>
-      val reservationId = maybeId.getOrElse(Reservation.inferReservationId(tasksMap, instanceId))
-      Reservation(volumesIds, state, reservationId)
-    }
-  }
-
   /**
     * Read format for old instance with created condition.
     */
@@ -113,7 +102,7 @@ object MigrationTo17 extends StrictLogging {
       (__ \ "reservation").readNullable[JsObject]
     ) { (instanceId, agentInfo, tasksMap, runSpecVersion, state, rawReservation) =>
         val reservation = rawReservation.map { raw =>
-          raw.as[Reservation](legacyReservationReads(tasksMap, instanceId))
+          raw.as[Reservation](InstanceMigration.legacyReservationReads(tasksMap, instanceId))
         }
         new Instance(instanceId, Some(agentInfo), state, tasksMap, runSpecVersion, reservation)
       }
